refactor(kheavyhash): apply review feedback (uint found, drop meta-comments, notFindNonce tests)
- kheavyhash_result.cl: drop header comment and MAX_RESULT fallback define
  (the value is always supplied by the kernel generator via addDefine);
  widen found from uchar to uint to remove implicit struct padding so the
  host and device Result layouts match exactly; put inline on its own line.
- result.hpp: widen Result::found from bool to uint32_t to mirror the device
  uint (kept uint32_t rather than cl_uint because this header is shared with
  the CUDA build, which has no OpenCL headers).
- opencl_kat_test.cpp: mirror the found widening in the KAT Result struct.
- kheavyhash.cl: drop the appended-by-resolver meta-comments.
- amd/nvidia tests: add notFindNonce coverage (minimum target -> no submit).

+++ sources/algo/kheavyhash/opencl/tests/opencl_kat_test.cpp
@@ -248,10 +248,10 @@ TEST_F(OpenClKat, heavyHashMatchesReference)
 class SearchKernel : public OpenClKat, public ::testing::WithParamInterface<char const*>
 {
   protected:
-    // Mirrors the kernel's Result struct (and algo::ethash::Result layout).
+    // Mirrors the kernel's Result struct (algo::kheavyhash::Result layout).
     struct alignas(8) Result
     {
-        uint8_t  found{ 0u };
+        uint32_t found{ 0u };
         uint32_t count{ 0u };
         uint64_t nonces[algo::kheavyhash::MAX_RESULT]{ 0ull, 0ull, 0ull, 0ull };
     };
